[FrontendSpec] Adding TTLCache to `try_get_grafana_service_url` function (#5543)
