+++ server/api/api/utils.py
@@ -15,10 +15,8 @@
 import asyncio
 import collections
 import copy
-import functools
 import json
 import re
-import time
 import traceback
 import typing
 import uuid
@@ -65,40 +63,6 @@ from server.api.utils.singletons.scheduler import get_scheduler
 def log_and_raise(status=HTTPStatus.BAD_REQUEST.value, **kw):
     logger.error(str(kw))
     raise HTTPException(status_code=status, detail=kw)
-
-
-def lru_cache_with_ttl(maxsize=128, typed=False, ttl_seconds=60):
-    """
-    Thread-safety least-recently used cache with time-to-live (ttl_seconds) limit.
-    https://stackoverflow.com/a/71634221/5257501
-    """
-
-    class Result:
-        __slots__ = ("value", "death")
-
-        def __init__(self, value, death):
-            self.value = value
-            self.death = death
-
-    def decorator(func):
-        @functools.lru_cache(maxsize=maxsize, typed=typed)
-        def cached_func(*args, **kwargs):
-            value = func(*args, **kwargs)
-            death = time.monotonic() + ttl_seconds
-            return Result(value, death)
-
-        @functools.wraps(func)
-        def wrapper(*args, **kwargs):
-            result = cached_func(*args, **kwargs)
-            if result.death < time.monotonic():
-                result.value = func(*args, **kwargs)
-                result.death = time.monotonic() + ttl_seconds
-            return result.value
-
-        wrapper.cache_clear = cached_func.cache_clear
-        return wrapper
-
-    return decorator
 
 
 def log_path(project, uid) -> Path:
